+++ src/CookTime/client-app/src/components/Tags/TagList.tsx
@@ -8,7 +8,7 @@ type TagListProps = {
 // Ostensibly this is just the tag list
 export class TagList extends React.Component<TagListProps, {}> {
     render() {
-        let lis = this.props.data.map((item, i) => {
+        const lis = this.props.data.map((item, i) => {
             return (<span key={i} className="badge">
                 <input type="hidden" name={`Skills[${i}]`} value={item}></input>
                 {item}